more refactors + discretizing the distance
