+++ commons/servers/BUILD.bazel
@@ -3,8 +3,8 @@ load("@npm_bazel_typescript//:index.bzl", "ts_library")
 package(default_visibility = ["//visibility:public"])
 
 ts_library(
-    name = "commons",
-    srcs = glob(["src/**/*.ts"]),
+    name = "servers",
+    srcs = glob(["*.ts"]),
     module_name = "decentraland-katalyst-commons",
     tsconfig = "//:tsconfig.json",
     deps = [
