+++ commons/utils/BUILD.bazel
@@ -1,0 +1,10 @@
+load("@npm_bazel_typescript//:index.bzl", "ts_library")
+
+package(default_visibility = ["//visibility:public"])
+
+ts_library(
+    name = "utils",
+    srcs = glob(["*.ts"]),
+    module_name = "decentraland-katalyst-utils",
+    tsconfig = "//:tsconfig.json",
+)
