+++ comms/lighthouse/BUILD.bazel
@@ -62,7 +62,8 @@ ts_library(
     srcs = glob(["src/**/*.ts"]),
     module_name = "@katalyst/lighthouse",
     deps = [
-        "//commons",
+        "//commons/servers",
+        "//commons/utils",
         "//contracts",
         "@npm//@types",
         "@npm//cors",
@@ -107,7 +108,8 @@ ts_library(
     tsconfig = "//:tsconfig.json",
     deps = [
         ":lighthouse",
-        "//commons",
+        "//commons/servers",
+        "//commons/utils",
         "@npm//@types/jasmine",
     ],
 )
