+++ comms/lighthouse/src/naming.ts
@@ -1,6 +1,6 @@
-import { DAOClient } from "decentraland-katalyst-commons/src/DAOClient";
-import { ServerMetadata } from "decentraland-katalyst-commons/src/ServerMetadata";
-import { noReject } from "decentraland-katalyst-commons/src/util";
+import { DAOClient } from "decentraland-katalyst-commons/DAOClient";
+import { ServerMetadata } from "decentraland-katalyst-commons/ServerMetadata";
+import { noReject } from "decentraland-katalyst-utils/util";
 import { lighthouseStorage } from "./simpleStorage";
 
 export const defaultNames = ["zeus", "poseidon", "athena", "hera", "hephaestus", "aphrodite", "hades", "hermes", "artemis", "thor", "loki", "odin", "freyja", "fenrir", "heimdallr", "baldr"];
