+++ comms/lighthouse/test/naming.spec.ts
@@ -1,9 +1,9 @@
 // Because of Bazel sandboxing, we need this for the time being
 process.env.LIGHTHOUSE_STORAGE_LOCATION = ".";
 
-import { DAOClient } from "decentraland-katalyst-commons/src/DAOClient";
+import { DAOClient } from "decentraland-katalyst-commons/DAOClient";
 import { pickName, defaultNames } from "../src/naming";
-import { ServerMetadata } from "decentraland-katalyst-commons/src/ServerMetadata";
+import { ServerMetadata } from "decentraland-katalyst-commons/ServerMetadata";
 import { lighthouseStorage } from "../src/simpleStorage";
 
 declare var global: any;
