+++ content/BUILD.bazel
@@ -61,7 +61,8 @@ ts_library(
     srcs = glob(["test/integration/**/*.ts"]),
     tsconfig = "//:tsconfig.json",
     deps = [
-        "//commons",
+        "//commons/servers",
+        "//commons/utils",
         "//content/src:content",
         "//content/test/helpers:tests_helpers",
         "@npm//@types/faker",
