+++ tsconfig.json
@@ -34,7 +34,8 @@
       "@katalyst/content/*": ["content/src/*"],
       "@katalyst/test-helpers/*": ["content/test/helpers/*"],
       "decentraland-katalyst-contracts/*": ["contracts/*"],
-      "decentraland-katalyst-commons/*": ["commons/*"],
+      "decentraland-katalyst-commons/*": ["commons/servers/*"],
+      "decentraland-katalyst-utils/*": ["commons/utils/*"],
       "katalyst/comms/peer/proto/*": ["dist/bin/comms/peer/proto/*"]
     }
   },
